Added a protection around a '[ -r xx ]' test. This fails with a file-not-found error on some systems??
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/fixincl-branch@24885 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/fixinc/inclhack.tpl
@@ -338,18 +338,20 @@ while [ $# != 0 ]; do
     #
     cd ${INPUT}
     cd $1
-    if [ -r $2 ] && [ ! -r $3 ]; then
-      cp $2 $3 >/dev/null 2>&1 || echo "Can't copy $2" >&2
-      chmod +w $3 2>/dev/null
-      chmod a+r $3 2>/dev/null
-      echo Copied $2
-      for include in `egrep '^[ 	]*#[ 	]*include[ 	]*"[^/]' $3 |
+    if [ -f $2 ] ; then
+      if [ -r $2 ] && [ ! -r $3 ]; then
+        cp $2 $3 >/dev/null 2>&1 || echo "Can't copy $2" >&2
+        chmod +w $3 2>/dev/null
+        chmod a+r $3 2>/dev/null
+        echo Copied $2
+        for include in `egrep '^[ 	]*#[ 	]*include[ 	]*"[^/]' $3 |
              sed -e 's/^[ 	]*#[ 	]*include[ 	]*"\([^"]*\)".*$/\1/'`
-      do
-	dir=`echo $2 | sed -e s'|/[^/]*$||'`
-	dir2=`echo $3 | sed -e s'|/[^/]*$||'`
-	newreq="$newreq $1 $dir/$include $dir2/$include"
-      done
+        do
+	  dir=`echo $2 | sed -e s'|/[^/]*$||'`
+	  dir2=`echo $3 | sed -e s'|/[^/]*$||'`
+	  newreq="$newreq $1 $dir/$include $dir2/$include"
+        done
+      fi
     fi
     shift; shift; shift
   done
